requirements.txt
# Core dependencies for theta-alpha shift analysis
mne>=1.6
neurodsp>=2.2
specparam>=1.1
bycycle>=1.1
emd>=0.6
numpy>=1.24
scipy>=1.11
pandas>=2.0
matplotlib>=3.7
seaborn>=0.13
PyYAML>=6.0
tqdm>=4.65
pytest>=7.4
scikit-learn>=1.3

# Web application
flask>=3.0
gunicorn>=21.2

# Optional — heavy deps, install when needed
# alphacsc>=0.4
# hmmlearn>=0.3
